moved function to registration.utils
